Move lfilter and convolve into functional, more tests
