+++ torchaudio/functional_filtering.py
@@ -2,37 +2,9 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 import math
 import torch
 import torchaudio
-
-__all__ = ["lowpass_biquad", "highpass_biquad", "biquad", "lfilter", "convolve"]
-
 from _torch_filtering import diff_eq as diffeq_cpp
 
-
-def convolve(input_waveform, impulse_response):
-
-    n_channels, n_frames = input_waveform.size()
-
-    # for each audio channel
-    assert input_waveform.dtype == torch.float32
-    assert impulse_response.dtype == torch.float32
-
-    channel_output_waveforms = []
-    for i_channel in range(n_channels):
-        channel_output_waveforms.append(
-            torch.nn.functional.conv_transpose1d(
-                input_waveform[i_channel, :].unsqueeze(0).unsqueeze(0),
-                impulse_response.unsqueeze(0).unsqueeze(0),
-            ).squeeze(0).squeeze(0)
-        )
-    return torch.stack(channel_output_waveforms)
-
-
-def lfilter(input_waveform, a_coeffs, b_coeffs):
-    output_waveform = torch.zeros_like(input_waveform)
-    assert input_waveform.dtype == torch.float32
-
-    diffeq_cpp(input_waveform, output_waveform, a_coeffs, b_coeffs)
-    return output_waveform
+__all__ = ["lowpass_biquad", "highpass_biquad", "biquad"]
 
 
 def biquad(input_waveform, b0, b1, b2, a0, a1, a2):
